libft ready

+++ libft/ft_strmapi.c
@@ -3,14 +3,15 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strmapi.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: vpozniak <vpozniak@student.42warsaw.pl>    +#+  +:+       +#+        */
+/*   By: tafanasi <tafanasi@student.42warsaw.pl>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/12/15 18:26:27 by vpozniak          #+#    #+#             */
-/*   Updated: 2024/12/15 18:30:21 by vpozniak         ###   ########.fr       */
+/*   Created: 2024/12/16 17:11:05 by tafanasi          #+#    #+#             */
+/*   Updated: 2024/12/16 17:11:06 by tafanasi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
+#include <stdlib.h>
 
 char	*ft_strmapi(char const *s, char (*f)(unsigned int, char))
 {
@@ -21,7 +22,7 @@ char	*ft_strmapi(char const *s, char (*f)(unsigned int, char))
 	if (!res)
 		return (NULL);
 	i = 0;
-	while (i < ft_strlen(s))
+	while (i < (unsigned int)ft_strlen(s))
 	{
 		res[i] = (*f)(i, s[i]);
 		i++;
